Move EventScheduleItem into the same dir as EventSchedule
