+++ common/views/components/EventSchedule/EventScheduleItem.js
